refactor: split tools into one file per tool
Each tool in agents/tools/ is now its own class implementing Tool<TArgs,
TResult> with definition() and call() methods, matching the pattern from
the reference project. Previously tools were grouped (EditorTools.ts,
WorkspaceTools.ts, DelegationTools.ts each contained multiple tools).

- 23 new individual tool class files (read, edit, write, invoke_agent, etc.)
- EditorTools/WorkspaceTools retain only their state-holder classes
- DelegationTools.ts becomes a thin registrar over individual classes
- registries.ts imports and registers individual tool classes directly
- AgentContext.tsx uses DelegateToSkillTool class instead of the old handler factory
- delegate_to_skill tests moved to delegate_to_skill.test.ts
