+++ src/lib/agents/tools/EditorTools.ts
@@ -2,16 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import * as monaco from "monaco-editor";
-import {
-  AgentConfig,
-  AgentEvent,
-  ToolContext,
-  ToolProvider,
-  ToolRegistry,
-} from "@mast-ai/core";
 import { Suggestion } from "../../store";
-import type { AgentRunnerFactory } from "../";
-import { loadSkills } from "../../skills";
 import { v4 as uuidv4 } from "uuid";
 
 export class EditorTools {
@@ -192,175 +183,4 @@ export class EditorTools {
       this.setSuggestions((prev) => [...prev, newSuggestion]);
     });
   }
-}
-
-/** Registers the standard editor tools on a ToolRegistry. */
-export function registerEditorTools(
-  registry: ToolRegistry,
-  tools: EditorTools,
-): void {
-  registry.register({
-    definition: () => ({
-      name: "read",
-      description: "Reads the complete current editor content.",
-      parameters: { type: "object", properties: {} },
-      scope: "read" as const,
-    }),
-    call: async () => tools.read(),
-  });
-
-  registry.register({
-    definition: () => ({
-      name: "read_selection",
-      description: "Reads the currently selected text in the editor.",
-      parameters: { type: "object", properties: {} },
-      scope: "read" as const,
-    }),
-    call: async () => tools.read_selection(),
-  });
-
-  registry.register({
-    definition: () => ({
-      name: "search",
-      description:
-        "Finds all occurrences of a query string in the document. Returns the line and column of each match.",
-      parameters: {
-        type: "object",
-        properties: {
-          query: { type: "string", description: "The text to search for." },
-        },
-        required: ["query"],
-      },
-      scope: "read" as const,
-    }),
-    call: async (args: { query: string }) => tools.search(args),
-  });
-
-  registry.register({
-    definition: () => ({
-      name: "get_metadata",
-      description:
-        "Returns metadata about the current document: character count, word count, and line count.",
-      parameters: { type: "object", properties: {} },
-      scope: "read" as const,
-    }),
-    call: async () => tools.get_metadata(),
-  });
-
-  registry.register({
-    definition: () => ({
-      name: "get_current_mode",
-      description:
-        "Returns the current UI mode: 'editor' (Monaco editor is visible) or 'preview' (Markdown preview is visible). Check this before making edits to ensure the editor is accessible.",
-      parameters: { type: "object", properties: {} },
-      scope: "read" as const,
-    }),
-    call: async () => tools.get_current_mode(),
-  });
-
-  registry.register({
-    definition: () => ({
-      name: "edit",
-      description:
-        "Proposes a targeted edit. This tool pauses and waits for user approval. ONLY use this for small, localized changes (e.g., 1-2 sentences). Never pass the entire document.",
-      parameters: {
-        type: "object",
-        properties: {
-          originalText: {
-            type: "string",
-            description:
-              "The exact, minimal string of text to replace. Must be short. Do NOT pass the whole document.",
-          },
-          replacementText: {
-            type: "string",
-            description: "The new text to replace the originalText with.",
-          },
-        },
-        required: ["originalText", "replacementText"],
-      },
-      scope: "write" as const,
-    }),
-    call: async (args: { originalText: string; replacementText: string }) =>
-      tools.edit(args),
-  });
-
-  registry.register({
-    definition: () => ({
-      name: "write",
-      description:
-        "Proposes a complete rewrite. This tool pauses and waits for user approval. ONLY use this when the user explicitly requests a total rewrite of the entire document.",
-      parameters: {
-        type: "object",
-        properties: {
-          content: {
-            type: "string",
-            description: "The full new document content.",
-          },
-        },
-        required: ["content"],
-      },
-      scope: "write" as const,
-    }),
-    call: async (args: { content: string }) => tools.write(args),
-  });
-
-  registry.register({
-    definition: () => ({
-      name: "request_switch_to_editor",
-      description:
-        "Requests the user to switch from Preview mode to Editor mode. This will display a prompt to the user and pause until they accept or decline. Call this before attempting edits when in preview mode.",
-      parameters: { type: "object", properties: {} },
-      scope: "write" as const,
-    }),
-    call: async () => tools.request_switch_to_editor(),
-  });
-}
-
-/**
- * Creates the delegate_to_skill tool call handler.
- * Extracted for testability; the factory parameter can be overridden in tests.
- */
-type RunnerLike = {
-  runBuilder: (agent: AgentConfig) => {
-    runStream: (input: string) => AsyncIterable<AgentEvent>;
-  };
-};
-
-export function createDelegateToSkillHandler(
-  factory: AgentRunnerFactory,
-  readonlyRegistry: ToolProvider,
-  runnerFactory: (registry: ToolProvider, model?: string) => RunnerLike = (
-    registry,
-    model,
-  ) => factory.create({ tools: registry, model }),
-): (
-  args: { skillName: string; task: string },
-  context: ToolContext,
-) => Promise<string> {
-  return async ({ skillName, task }, context) => {
-    const skills = loadSkills();
-    const skill = skills.find((s) => s.name === skillName);
-    if (!skill) {
-      const names = skills.map((s) => s.name).join(", ");
-      return `Error: skill "${skillName}" not found. Available skills: ${names || "none"}`;
-    }
-
-    const readonlyToolNames = readonlyRegistry.getTools().map((d) => d.name);
-
-    const childRunner = runnerFactory(readonlyRegistry, skill.model);
-    const agentConfig: AgentConfig = {
-      name: skill.name,
-      instructions: skill.instructions,
-      tools: readonlyToolNames,
-    };
-    for await (const event of childRunner
-      .runBuilder(agentConfig)
-      .runStream(task)) {
-      if (event.type === "done") {
-        return event.output;
-      }
-      context.onEvent?.(event);
-    }
-    throw new Error("Child runner ended without a done event");
-  };
 }
