+++ src/lib/agents/tools/WorkspaceTools.ts
@@ -1,7 +1,7 @@
 // Copyright 2026 Andre Cipriani Bandarra
 // SPDX-License-Identifier: Apache-2.0
 
-import { AgentConfig, ToolRegistry } from "@mast-ai/core";
+import { AgentConfig } from "@mast-ai/core";
 import { v4 as uuidv4 } from "uuid";
 import { WorkspaceDocument } from "../../workspace";
 import type { WorkspaceActionRequest } from "../../store";
@@ -205,181 +205,4 @@ export class WorkspaceTools {
     const result = await runResearch(query, this.docsRef.current, this.factory);
     return JSON.stringify(result);
   }
-}
-
-export function registerWorkspaceTools(
-  registry: ToolRegistry,
-  tools: WorkspaceTools,
-): void {
-  registry.register({
-    definition: () => ({
-      name: "get_active_doc_info",
-      description:
-        "Returns the id and title of the document currently open in the editor.",
-      parameters: { type: "object", properties: {} },
-      scope: "read" as const,
-    }),
-    call: async () => tools.get_active_doc_info(),
-  });
-
-  registry.register({
-    definition: () => ({
-      name: "list_workspace_docs",
-      description:
-        "Lists all documents in the workspace. Returns an array of { id, title } objects.",
-      parameters: { type: "object", properties: {} },
-      scope: "read" as const,
-    }),
-    call: async () => tools.list_workspace_docs(),
-  });
-
-  registry.register({
-    definition: () => ({
-      name: "read_workspace_doc",
-      description:
-        "Reads the full content of a specific document. Returns { title, content } or { error }.",
-      parameters: {
-        type: "object",
-        properties: {
-          id: { type: "string", description: "The document ID to read." },
-        },
-        required: ["id"],
-      },
-      scope: "read" as const,
-    }),
-    call: async (args: { id: string }) => tools.read_workspace_doc(args),
-  });
-
-  registry.register({
-    definition: () => ({
-      name: "query_workspace_doc",
-      description:
-        "Asks a question about a specific document using a sub-agent. Returns { summary, excerpt } where excerpt is the most relevant verbatim passage.",
-      parameters: {
-        type: "object",
-        properties: {
-          id: { type: "string", description: "The document ID to query." },
-          query: {
-            type: "string",
-            description: "The question about the document.",
-          },
-        },
-        required: ["id", "query"],
-      },
-      scope: "read" as const,
-    }),
-    call: async (args: { id: string; query: string }) =>
-      tools.query_workspace_doc(args),
-  });
-
-  registry.register({
-    definition: () => ({
-      name: "query_workspace",
-      description:
-        "Asks a question spanning all workspace documents and synthesizes the results. Returns { summary, sources: [{ id, title, excerpt }] }.",
-      parameters: {
-        type: "object",
-        properties: {
-          query: {
-            type: "string",
-            description: "The question to answer across all documents.",
-          },
-        },
-        required: ["query"],
-      },
-      scope: "read" as const,
-    }),
-    call: async (args: { query: string }) => tools.query_workspace(args),
-  });
-
-  registry.register({
-    definition: () => ({
-      name: "create_document",
-      description:
-        "Creates a new document in the workspace with the given title and optional initial content. Providing content avoids a separate write step. Pauses for user authorization before creating.",
-      parameters: {
-        type: "object",
-        properties: {
-          title: {
-            type: "string",
-            description: "The title for the new document.",
-          },
-          content: {
-            type: "string",
-            description:
-              "Optional initial content for the new document. If omitted the document is created blank.",
-          },
-        },
-        required: ["title"],
-      },
-      scope: "write" as const,
-    }),
-    call: async (args: { title: string; content?: string }) =>
-      tools.create_document(args),
-  });
-
-  registry.register({
-    definition: () => ({
-      name: "rename_document",
-      description:
-        "Renames an existing document in the workspace. Pauses for user authorization before renaming.",
-      parameters: {
-        type: "object",
-        properties: {
-          id: {
-            type: "string",
-            description: "The document ID to rename.",
-          },
-          title: {
-            type: "string",
-            description: "The new title for the document.",
-          },
-        },
-        required: ["id", "title"],
-      },
-      scope: "write" as const,
-    }),
-    call: async (args: { id: string; title: string }) =>
-      tools.rename_document(args),
-  });
-
-  registry.register({
-    definition: () => ({
-      name: "delete_document",
-      description:
-        "Deletes a document from the workspace. Pauses for user authorization before deleting. If the deleted document was active, a different document becomes active.",
-      parameters: {
-        type: "object",
-        properties: {
-          id: {
-            type: "string",
-            description: "The document ID to delete.",
-          },
-        },
-        required: ["id"],
-      },
-      scope: "write" as const,
-    }),
-    call: async (args: { id: string }) => tools.delete_document(args),
-  });
-
-  registry.register({
-    definition: () => ({
-      name: "switch_active_document",
-      description:
-        "Switches the active document in the editor. Saves the current document content before switching. Does not require user authorization.",
-      parameters: {
-        type: "object",
-        properties: {
-          id: {
-            type: "string",
-            description: "The document ID to switch to.",
-          },
-        },
-        required: ["id"],
-      },
-      scope: "write" as const,
-    }),
-    call: async (args: { id: string }) => tools.switch_active_document(args),
-  });
 }
